backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
pymongo==4.5.0
flask-cors==4.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
gunicorn==21.2.0
marshmallow==3.20.1
